pyproject.toml
[tool.poetry]
name = "code_graph"
version = "0.1.0"
description = "code_graph is designed to help developers visualize and analyze the structure of their source code. It takes source code as input and generates a graph representation, making it easier to understand relationships and dependencies within the codebase."
authors = ["Roi Lipman <roilipman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
tree-sitter = "^0.22.3"
GitPython = "^3.1.43"
falkordb = "^1.0.5"
tree-sitter-c = "^0.21.4"
tree-sitter-python = "^0.21.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
